+++ tests/ingest/test_models.py
@@ -1,11 +1,10 @@
-from mex.backend.ingest.models import _BaseBulkIngestRequest
+from mex.backend.ingest.models import BulkIngestRequest
 
 
-class DummyBulkIngestRequest(_BaseBulkIngestRequest):
-    things: list[str]
-    stuff: list[str]
+class DummyBulkIngestRequest(BulkIngestRequest):
+    items: list[str]
 
 
 def test_bulk_ingest_request_get_all() -> None:
-    dummy_request = DummyBulkIngestRequest(things=["a", "b", "c"], stuff=["d"])
-    assert dummy_request.get_all() == ["a", "b", "c", "d"]
+    dummy_request = DummyBulkIngestRequest(items=["a", "b", "c"])
+    assert dummy_request.items == ["a", "b", "c"]