+++ CHANGELOG.md
@@ -15,10 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - BREAKING: make `MEX_EXTRACTED_PRIMARY_SOURCE` an instance of its own class
   instead of ExtractedPrimarySource in order to set static provenance identifiers
-
+- BulkIngestRequest contains now one single list "items"
+- tests for ingestion adapted to BulkIngestRequest-Model
 ### Deprecated
 
 ### Removed
+- removed class  _BaseBulkIngestRequest for ingestion model
 
 ### Fixed